punched up service layer tests for comments and likes

+++ PythonAPI/service_layer/implementation_classes/comment_service.py
@@ -1,5 +1,10 @@
 from typing import List
 
+from custom_exceptions.comment_id_must_be_an_integer import CommentIdMustBeAnInteger
+from custom_exceptions.comment_must_be_a_string import CommentMustBeAString
+from custom_exceptions.group_id_must_be_an_integer import GroupIdMustBeAnInteger
+from custom_exceptions.post_id_must_be_an_integer import PostIdMustBeAnInteger
+from custom_exceptions.user_id_must_be_an_integer import UserIdMustBeAnInteger
 from data_access_layer.implementation_classes.comment_dao import CommentDAOImp
 from entities.comment import Comment
 from entities.returned_comment import ReturnedComment
@@ -12,10 +17,35 @@ class CommentServiceImp(CommentService):
 
     def service_create_comment(self, post_id: int, user_id: int, comment_text: str, group_id: int,
                                reply_user: int) -> Comment:
+        # Check to make sure the user_id is an integer
+        if not str(user_id).isnumeric():
+            raise UserIdMustBeAnInteger('The user id must be an integer.')
+
+        # Check to make sure the post_id is an integer
+        if not str(post_id).isnumeric():
+            raise PostIdMustBeAnInteger('The post id must be an integer.')
+
+        # Check to make sure the group_id is an integer
+        if not str(group_id).isnumeric():
+            raise GroupIdMustBeAnInteger('The group id must be an integer.')
+
+        # Check to make sure that the image format is a string
+        if not type(comment_text) == str:
+            raise CommentMustBeAString('The comment must be a string.')
+
         return self.comment_dao.create_comment(post_id, user_id, comment_text, group_id, reply_user)
 
     def service_get_comment_by_post_id(self, post_id: int) -> List[ReturnedComment]:
+
+        # Check to make sure the post_id is an integer
+        if not str(post_id).isnumeric():
+            raise PostIdMustBeAnInteger('The post id must be an integer.')
+
         return self.comment_dao.get_comment_by_post_id(post_id)
 
     def service_delete_comment(self, comment_id: int) -> bool:
+        # Check to make sure the post_id is an integer
+        if not str(comment_id).isnumeric():
+            raise CommentIdMustBeAnInteger('The comment id must be an integer.')
+
         return self.comment_dao.delete_comment(comment_id)
